secrets/pyproject.toml
[tool.poetry]
name = "pgpemu-esp32"
version = "0.1.0"
description = ""
authors = ["spezifisch <spezifisch23@proton.me>"]
readme = "README.md"
packages = [{include = "pgpemu_esp32"}]

[tool.poetry.dependencies]
python = "^3.9"
pyserial = "^3.5"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

